refactor: moving add database form utils into separate class
